requirements.txt
streamlit
requests
Pillow
tweepy
python-dotenv
openai
snowflake-connector-python
pandas
plotly